move pausemenu to separate folder, add customization options
